nbackend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
firebase-admin==6.2.0

# PDF Processing - UPDATED
PyPDF2==3.0.1
pdfplumber==0.10.3

# AI/ML - UPDATED: Removed transformers and torch
google-generativeai==0.3.2

# File Export
reportlab==4.0.7
python-docx==1.1.0

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
python-jose[cryptography]==3.3.0